+++ Specs/DynamicScene/DynamicBillboardSpec.js
@@ -3,6 +3,7 @@ defineSuite([
              'DynamicScene/DynamicBillboard',
              'Core/Cartesian2',
              'Core/Cartesian3',
+             'Core/NearFarScalar',
              'Core/Color',
              'Scene/HorizontalOrigin',
              'Scene/VerticalOrigin',
@@ -11,6 +12,7 @@ defineSuite([
              DynamicBillboard,
              Cartesian2,
              Cartesian3,
+             NearFarScalar,
              Color,
              HorizontalOrigin,
              VerticalOrigin,
@@ -30,6 +32,9 @@ defineSuite([
         source.pixelOffset = new ConstantProperty(Cartesian2.UNIT_X);
         source.scale = new ConstantProperty(1);
         source.show = new ConstantProperty(false);
+        source.width = new ConstantProperty(24);
+        source.height = new ConstantProperty(36);
+        source.nearFarScalar = new ConstantProperty(new NearFarScalar());
 
         var target = new DynamicBillboard();
         target.merge(source);
@@ -44,6 +49,9 @@ defineSuite([
         expect(target.pixelOffset).toBe(source.pixelOffset);
         expect(target.scale).toBe(source.scale);
         expect(target.show).toBe(source.show);
+        expect(target.width).toBe(source.width);
+        expect(target.height).toBe(source.height);
+        expect(target.nearFarScalar).toBe(source.nearFarScalar);
     });
 
     it('merge does not assign assigned properties', function() {
@@ -58,6 +66,9 @@ defineSuite([
         source.pixelOffset = new ConstantProperty(Cartesian2.UNIT_X);
         source.scale = new ConstantProperty(1);
         source.show = new ConstantProperty(false);
+        source.width = new ConstantProperty(24);
+        source.height = new ConstantProperty(36);
+        source.nearFarScalar = new ConstantProperty(new NearFarScalar());
 
         var image = new ConstantProperty('');
         var rotation = new ConstantProperty(5);
@@ -69,6 +80,9 @@ defineSuite([
         var pixelOffset = new ConstantProperty(Cartesian2.UNIT_X);
         var scale = new ConstantProperty(1);
         var show = new ConstantProperty(false);
+        var width = new ConstantProperty(2);
+        var height = new ConstantProperty(3);
+        var nearFarScalar = new ConstantProperty(new NearFarScalar());
 
         var target = new DynamicBillboard();
         target.image = image;
@@ -81,6 +95,9 @@ defineSuite([
         target.pixelOffset = pixelOffset;
         target.scale = scale;
         target.show = show;
+        target.width = width;
+        target.height = height;
+        target.nearFarScalar = nearFarScalar;
 
         target.merge(source);
 
@@ -94,6 +111,9 @@ defineSuite([
         expect(target.pixelOffset).toBe(pixelOffset);
         expect(target.scale).toBe(scale);
         expect(target.show).toBe(show);
+        expect(target.width).toBe(width);
+        expect(target.height).toBe(height);
+        expect(target.nearFarScalar).toBe(nearFarScalar);
     });
 
     it('clone works', function() {
@@ -108,6 +128,9 @@ defineSuite([
         source.pixelOffset = new ConstantProperty(Cartesian2.UNIT_X);
         source.scale = new ConstantProperty(1);
         source.show = new ConstantProperty(false);
+        source.width = new ConstantProperty(24);
+        source.height = new ConstantProperty(36);
+        source.nearFarScalar = new ConstantProperty(new NearFarScalar());
 
         var result = source.clone();
         expect(result.image).toBe(source.image);
@@ -120,6 +143,9 @@ defineSuite([
         expect(result.pixelOffset).toBe(source.pixelOffset);
         expect(result.scale).toBe(source.scale);
         expect(result.show).toBe(source.show);
+        expect(result.width).toBe(source.width);
+        expect(result.height).toBe(source.height);
+        expect(result.nearFarScalar).toBe(source.nearFarScalar);
     });
 
     it('merge throws if source undefined', function() {